Unbreak the query modified by Events Calendar on edit screens
The Events Calendar makes the decision that tribe_events items
should appear in every tag query via a `parse_query` action. This
causes some issues on the edit screen in WordPress, as a string is
expected for the post_type data.

This change resets the query back to the originally requested
post type.