Only force remove events from page list in main query
The Events Calendar stomps its way into the page listing, so we
stomped back in 2409beceb360

Unfortunately, this stomped out our Last Updated column as well
since it searches for revisions and was being told to only use
pages all of a sudden.

Adding a check for is_main_query() ensures that we only stomp the
big one.

Fixes #35